requirements.txt
aiohttp==3.7.4.post0
alpha-vantage==2.3.1
async-timeout==3.0.1
attrs==21.2.0
Brotli==1.0.9
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dash==1.21.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
mysql==0.0.3
mysqlclient==2.0.3
numpy==1.21.1
pandas==1.3.1
plotly==5.2.1
PyMySQL==1.0.2
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
tenacity==8.0.1
typing-extensions==3.10.0.0
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.1
yarl==1.6.3
